requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
crossrefapi==1.4.0
entrypoints==0.3
flasgger==0.9.4
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-annotations==2.4.0
mccabe==0.6.1
mistune==0.8.4
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
six==1.14.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.25.7
uWSGI==2.0.18
Werkzeug==0.16.0
